Moved function getDateFromTimeSelection to timeMath
