+++ packages/date-time-utilities/etc/date-time-utilities.api.md
@@ -105,6 +105,9 @@ export const formatYear: (date: Date) => string;
 export const getBoundedDateRange: (dateRange: Date[], minDate?: Date | undefined, maxDate?: Date | undefined) => Date[];
 
 // @public
+export const getDateFromTimeSelection: (showSeconds: boolean, useHour12: boolean, baseDate: Date, selectedTime: string) => Date;
+
+// @public
 export function getDatePartHashValue(date: Date): number;
 
 // @public
